:recycle: change city id and dimension from int to size_t

+++ include/city.h
@@ -1,11 +1,12 @@
 #ifndef CITY_H
 #define CITY_H
 
+#include <stddef.h>
 #include <stdio.h>
 
 typedef struct city City;
 
-City *city_new(int id, double x, double y);
-int return_dimension(FILE *f);
+City *city_new(size_t id, double x, double y);
+size_t return_dimension(FILE *f);
 
 #endif // !CITY_H
